First iteration in moving plan.* code.
